Extract ImportDialogContainer
